Move IRBuilderTest definition to conftest
